Add files via upload
a new-age insurance company and employ
mutiple outreach plans to sell term insurance to your
customers. Telephonic marketing campaigns still remain one of
the most effective way to reach out to people however they
incur a lot of cost. Hence, it is important to identify the
customers that are most likely to convert beforehand so that
they can be specifically targeted via call. We are given the
historical marketing data of the insurance company and are
required to build a ML model that will predict if a client will
subscribe to the insurance